+++ tests/unit/builtin_monitors/kubernetes_monitor_test.py
@@ -22,6 +22,7 @@ import re
 import time
 import warnings
 import platform
+import tempfile
 from string import Template
 from collections import OrderedDict
 
@@ -57,11 +58,14 @@ from scalyr_agent.test_base import ScalyrTestCase, BaseScalyrLogCaptureTestCase
 from scalyr_agent.test_util import ScalyrTestUtils
 
 from tests.unit.monitor_utils.k8s_test import FakeCache
+from tests.unit.monitor_utils.k8s_test import KubernetesApi
 from tests.unit.configuration_test import TestConfigurationBase
 from tests.unit.copying_manager_tests.copying_manager_test import FakeMonitor
 from scalyr_agent.test_base import skipIf
 
 import mock
+import requests
+import requests_mock
 from mock import patch
 from six.moves import range
 
@@ -1267,3 +1271,195 @@ class CRIEnumeratorTestCase(TestConfigurationBase, ScalyrTestCase):
             result["cont-1"]["k8s_info"]["pod_info"].annotations,
             {"log.config.scalyr.com/attributes.parser": "test-5"},
         )
+
+
+class TestKubernetesKubeletApiAuthTokenCaching(TestConfigurationBase, ScalyrTestCase):
+    @skipIf(platform.system() == "Windows", "Skipping Linux only tests on Windows")
+    def test_re_read_token_every_x_seconds(self):
+        _, token_file_path = tempfile.mkstemp()
+
+        with open(token_file_path, "w") as fp:
+            fp.write("token1")
+
+        self._write_file_with_separator_conversion(
+            """ {
+            api_key: "hi there",
+            k8s_service_account_token: "%s"
+            k8s_token_re_read_interval: 1,
+          }
+        """
+            % (token_file_path)
+        )
+        global_config = self._create_test_configuration_instance()
+        global_config.parse()
+
+        k8s = KubernetesApi.create_instance(global_config=global_config)
+
+        # 1. On initial read we should read first token
+        self.assertEqual(k8s.token, "token1")
+        self.assertEqual(k8s.token, "token1")
+        self.assertEqual(k8s.token, "token1")
+
+        # Update the token
+        with open(token_file_path, "w") as fp:
+            fp.write("token2")
+
+        self.assertEqual(k8s.token, "token1")
+
+        # 2. After the token has been updated, the code should eventually re-read a new token
+        time.sleep(1.2)
+        self.assertEqual(k8s.token, "token2")
+        self.assertEqual(k8s.token, "token2")
+
+    @skipIf(platform.system() == "Windows", "Skipping Linux only tests on Windows")
+    def test_kubernetes_api_re_read_token_every_x_seconds(self):
+        _, token_file_path = tempfile.mkstemp()
+
+        with open(token_file_path, "w") as fp:
+            fp.write("k8s-token1")
+
+        self._write_file_with_separator_conversion(
+            """ {
+            api_key: "hi there",
+            k8s_service_account_token: "%s"
+            k8s_token_re_read_interval: 1,
+          }
+        """
+            % (token_file_path)
+        )
+        global_config = self._create_test_configuration_instance()
+        global_config.parse()
+
+        session = requests.Session()
+        adapter = requests_mock.Adapter()
+        session.mount("mock://", adapter)
+
+        expected_headers_token = ""
+
+        def custom_request_matcher(request):
+            """
+            Custom matcher which verify headers contain the correct auth token.
+            """
+            return "Authorization" in request.headers and request.headers[
+                "Authorization"
+            ] == "Bearer %s" % (expected_headers_token)
+
+        adapter.register_uri(
+            "GET",
+            "mock://test.com/test?pretty=0",
+            text="{}",
+            additional_matcher=custom_request_matcher,
+        )
+
+        k8s_api_url = "mock://test.com"
+
+        k8s = KubernetesApi.create_instance(
+            global_config=global_config,
+            k8s_api_url=k8s_api_url,
+            verify_api_queries=False,
+        )
+        k8s._session = session
+
+        # Token is read lazily on first request
+        expected_headers_token = "k8s-token1"
+
+        self.assertIsNone(k8s._token)
+        k8s.query_api("/test")
+
+        # Verify token has been read on initial request
+        self.assertEqual(k8s._token, "k8s-token1")
+
+        # Write a new token and verify it will be read after the timeout has been reached
+        with open(token_file_path, "w") as fp:
+            fp.write("k8s-token2")
+
+        k8s.query_api("/test")
+        self.assertEqual(k8s._token, "k8s-token1")
+
+        time.sleep(1.2)
+        expected_headers_token = "k8s-token2"
+
+        k8s.query_api("/test")
+        self.assertEqual(k8s._token, "k8s-token2")
+
+    @skipIf(platform.system() == "Windows", "Skipping Linux only tests on Windows")
+    def test_kubelet_api_re_read_token_every_x_seconds(self):
+        _, token_file_path = tempfile.mkstemp()
+
+        with open(token_file_path, "w") as fp:
+            fp.write("kubelet-token1")
+
+        self._write_file_with_separator_conversion(
+            """ {
+            api_key: "hi there",
+            k8s_service_account_token: "%s"
+            k8s_token_re_read_interval: 1,
+          }
+        """
+            % (token_file_path)
+        )
+        global_config = self._create_test_configuration_instance()
+        global_config.parse()
+
+        session = requests.Session()
+        adapter = requests_mock.Adapter()
+        session.mount("mock://", adapter)
+
+        expected_headers_token = ""
+
+        def custom_request_matcher(request):
+            """
+            Custom matcher which verify headers contain the correct auth token.
+            """
+            return "Authorization" in request.headers and request.headers[
+                "Authorization"
+            ] == "Bearer %s" % (expected_headers_token)
+
+        adapter.register_uri(
+            "GET",
+            "mock://test.com:10250/test",
+            text="{}",
+            additional_matcher=custom_request_matcher,
+        )
+
+        k8s_api_url = "mock://test.com"
+
+        k8s = KubernetesApi.create_instance(
+            global_config=global_config,
+            k8s_api_url=k8s_api_url,
+            verify_api_queries=False,
+        )
+        k8s._session = session
+
+        kubelet_url_template = Template("mock://${host_ip}:10250")
+
+        kubelet = KubeletApi(
+            k8s=k8s,
+            host_ip="test.com",
+            node_name="node-1",
+            verify_https=False,
+            kubelet_url_template=kubelet_url_template,
+        )
+        kubelet._session = session
+
+        # Token is read lazily on first request
+        expected_headers_token = "kubelet-token1"
+
+        self.assertIsNone(k8s._token)
+        kubelet.query_api("/test")
+
+        # Verify token has been read on initial request
+        self.assertEqual(k8s._token, "kubelet-token1")
+
+        # Write a new token and verify it will be read after the timeout has been reached
+        with open(token_file_path, "w") as fp:
+            fp.write("kubelet-token2")
+
+        kubelet.query_api("/test")
+        self.assertEqual(k8s._token, "kubelet-token1")
+
+        time.sleep(1.2)
+        expected_headers_token = "kubelet-token2"
+
+        kubelet.query_api("/test")
+        self.assertEqual(k8s._token, "kubelet-token2")