+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Scalyr Agent 2 Changes By Release
 =================================
 
+## 2.1.30 "TBD" - April 28, 2022
+
+<!---
+Packaged by Arthur Kamalov <arthurk@sentinelone.com> on Apr 25, 2022 23:04 -0800
+--->
+
+Kubernetes:
+* Agent has been updated to periodically try to re-read Kubernetes authentication token value from ``/var/run/secrets/kubernetes.io/serviceaccount/token`` file on disk (every 5 minutes by default). This way agent also supports Kubernetes deployments where token files are periodically automatically refreshed / rotated (e.g. https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/#service-account-token-volume-projection).
+
 ## 2.1.29 "Auter" - Mar 15, 2022
 
 <!---